requirements.txt
autopep8
pydantic
pymongo
python-decouple
wheel